requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
gunicorn==21.2.0